Convert column types in the comparisonType callback instead of a hook
